+++ lib/advanced_value_notifier.dart
@@ -1,9 +1,12 @@
 library advanced_value_notifier;
 
-export 'src/notifiers/simple_notifier.dart';
-export 'src/builder/simple_builder.dart';
-export 'src/listener/simple_listener.dart';
+export 'src/notifiers/history_notifier.dart';
+export 'src/builder/history_value_listenable_builder.dart';
+export 'src/listener/history_value_listenable_listener.dart';
 
 export 'src/notifiers/transformer_notifier.dart';
-export 'src/builder/transformer_builder.dart';
-export 'src/listener/transformer_listener.dart';
+export 'src/builder/transformer_value_listenable_builder.dart';
+export 'src/listener/transformer_value_listenable_listener.dart';
+
+export 'src/builder/simple_value_listenable_builder.dart';
+export 'src/listener/simple_value_listenable_listener.dart';