+++ lib/src/condition/history.dart
@@ -1,0 +1,1 @@
+typedef HistoryValueCondition<T> = bool Function(T? prev, T curr);