+++ lib/src/condition/simple.dart
@@ -1,0 +1,1 @@
+typedef ValueCondition<T> = bool Function(T value);