+++ lib/src/condition/transformer.dart
@@ -1,0 +1,6 @@
+typedef TransformerHistoryValueCondition<T, U> = bool Function(
+  T? prevValue,
+  U? prevTransformedValue,
+  T curr,
+  U currTransformedValue,
+);