+++ lib/src/notifiers/transformer_notifier.dart
@@ -1,6 +1,8 @@
+import 'package:advanced_value_notifier/src/notifiers/transformer_history_value_listenable.dart';
 import 'package:flutter/foundation.dart';
 
-class TransformerHistoryValueNotifier<T, U> extends ChangeNotifier {
+class TransformerHistoryValueNotifier<T, U> extends ChangeNotifier
+    implements TransformerHistoryValueListenable<T, U> {
   TransformerHistoryValueNotifier({required T value, required this.transformer})
       : _value = value,
         _transformedValue = transformer(value) {
@@ -9,20 +11,26 @@ class TransformerHistoryValueNotifier<T, U> extends ChangeNotifier {
     }
   }
 
+  @override
   T get value => _value;
   T _value;
 
   U _transformedValue;
+  @override
   U get transformedValue => _transformedValue;
 
   T? _prevValue;
+  @override
   T? get prevValue => _prevValue;
 
   U? _prevTransformedValue;
+  @override
   U? get prevTransformedValue => _prevTransformedValue;
 
-  U Function(T value) transformer;
+  @override
+  final U Function(T value) transformer;
 
+  @override
   set value(T newValue) {
     if (_value == newValue) {
       return;
@@ -35,5 +43,7 @@ class TransformerHistoryValueNotifier<T, U> extends ChangeNotifier {
   }
 
   @override
-  String toString() => '${describeIdentity(this)}($prevValue)($value)';
+  String toString() => '${describeIdentity(this)}prevValue:$prevValue '
+      'prevTransformedValue:$prevTransformedValue value:$value '
+      'transformedValue:$transformedValue';
 }